+++ back/src/routers/commentRouter.js
@@ -1,7 +1,6 @@
 import { Router } from "express";
 import { loginRequired } from "../middlewares/loginRequired";
-import { CommentService } from "../services/commentService";
-
+import { commentsReadController, commentCreateController, commentDeleteController } from "../contoller/commentController"
 const commentRouter = Router();
 
 /**
@@ -34,20 +33,7 @@ const commentRouter = Router();
  *       "500":
  *          description: 서버 에러
  */
-commentRouter.get("/talks/:talkId/comments", async function (req, res, next) {
-  try {
-    const talkId = Number(req.params.talkId);
-
-    const comments = await CommentService.getComments(talkId);
-    if (comments.errorMessage) {
-      throw new Error(comments.errorMessage);
-    }
-
-    res.status(200).send(comments);
-  } catch (error) {
-    next(error);
-  }
-});
+commentRouter.get("/talks/:talkId/comments", commentsReadController);
 
 /**
  * @swagger
@@ -90,39 +76,7 @@ commentRouter.get("/talks/:talkId/comments", async function (req, res, next) {
  *       "500":
  *          description: 서버 에러
  */
-commentRouter.post(
-  "/comments/comment",
-  loginRequired,
-  async function (req, res, next) {
-    try {
-      const mode = req.body.mode;
-      const userId = req.currentUserId;
-      const talkId = Number(req.body.talkId);
-      const parentCommentId =
-        mode === "reply" ? req.body.parentCommentId : null;
-      const comment = req.body.comment;
-
-      let result;
-      if (mode === "comment") {
-        result = await CommentService.addComment(talkId, comment, userId);
-      } else if (mode === "reply") {
-        result = await CommentService.addReply(
-          parentCommentId,
-          comment,
-          userId
-        );
-      }
-
-      if (result.errorMessage) {
-        throw new Error(result.errorMessage);
-      }
-
-      res.status(201).send(result);
-    } catch (error) {
-      next(error);
-    }
-  }
-);
+commentRouter.post("/comments/comment", loginRequired, commentCreateController);
 
 /**
  * @swagger
@@ -158,31 +112,6 @@ commentRouter.post(
  *       "500":
  *          description: 서버 에러
  */
-commentRouter.delete(
-  "/comments/:comment_id",
-  loginRequired,
-  async function (req, res, next) {
-    try {
-      const mode = req.query.mode;
-      const userId = req.currentUserId;
-      const comment_id = req.params.comment_id;
-      
-      let result;
-      if (mode === "comment") {
-        result = await CommentService.deleteComment(comment_id, userId);
-      } else if (mode === "reply") {
-        result = await CommentService.deleteReply(comment_id, userId);
-      }
-
-      if (result.errorMessage) {
-        throw new Error(result.errorMessage);
-      }
-
-      res.status(200).send(result);
-    } catch (error) {
-      next(error);
-    }
-  }
-);
+commentRouter.delete("/comments/:comment_id", loginRequired, commentDeleteController);
 
 export { commentRouter };
